requirements.txt
annotated-types==0.7.0
certifi==2025.10.5
charset-normalizer==3.4.3
dnspython==2.8.0
email-validator==2.3.0
greenlet==3.2.4
idna==3.10
iniconfig==2.1.0
loguru==0.7.3
packaging==25.0
playwright==1.55.0
pluggy==1.6.0
pydantic==2.12.0
pydantic_core==2.41.1
pyee==13.0.0
Pygments==2.19.2
pytest==8.4.2
pytest-base-url==2.1.0
pytest-playwright==0.7.1
python-dotenv==1.1.1
python-slugify==8.0.4
requests==2.32.5
text-unidecode==1.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
